+++ src/WinTool/Views/IDialog.cs
@@ -1,0 +1,8 @@
+using WinTool.Models;
+
+namespace WinTool.Views;
+
+public interface IDialog<TIn, TOut>
+{
+    Result<TOut> ShowDialog(TIn input);
+}